// api/src/core/entities/tv.entity.ts
export class Tv {
  id: number;
  name: string;
  original_name: string;
  original_language: string;
  overview: string;
  first_air_date: string;
  genre_ids: number[];
  popularity: number;
  vote_average: number;
  vote_count: number;
  backdrop_path: string;
  poster_path: string;
  origin_country: string[];
}